Fix step 2 duration bugs and trim subtitle text
- Max hours reduced from 12 to 6; add contact-us note for longer events
- Floor current_hours to min_hours to prevent cross-attraction bleed
- Fix strict !== type comparison bug in step1 (session int vs PHP int)
- Remove hourly-rate note from subtitle; show only activity + min hours

+++ public/booking/step1.php
@@ -26,7 +26,8 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
 
         if ($attraction) {
             // If attraction changed, clear downstream wizard state
-            if (wizard_get('attraction_id') !== $attraction_id) {
+            // Cast both to int — sessions can return stored ints as strings
+            if ((int)wizard_get('attraction_id') !== $attraction_id) {
                 wizard_clear();
             }
             wizard_set('attraction_id',  $attraction_id);
